FEAT: rt파일 유효성검사 초안
