+++ include/parse.h
@@ -40,18 +40,39 @@ int			ft_str_append(t_string *t_str, char c);
 
 // parse
 t_error	validate_file(int argc, char **argv, t_rt_list **file);
+
+// is_valid_elements_utils_base
+t_is_return	is_valid_double_format(char *to_double_str);
+t_is_return is_between(double conditional_num1, double conditional_num2, double num);
+t_is_return is_xyz_element(char *str);
+t_is_return	is_rgb_element(char *str);
+t_is_return	is_vector_element(char *str);
+
+// is_valid_elements_utils
+t_is_return	is_xyz_coordinate(char *str);
+t_is_return	is_rgb(char *str);
+t_is_return	is_rgb(char *str);
+t_is_return	is_3d_norm_orient_vecotor(char *str);
+
+// is_valid_elements
+t_is_return	is_valid_a_element(char **elements);
+t_is_return	is_valid_c_element(char **elements);
+t_is_return	is_valid_l_element(char **elements);
+t_is_return	is_valid_sp_element(char **elements);
+t_is_return	is_valid_pl_element(char **elements);
+t_is_return is_valid_cy_element(char **elements);
+
+
+//validate_elements
+t_is_return	is_rt_file(char *file);
+t_is_return	is_valid_type(char *type);
+t_is_return	is_valid_elements(char **elements);
+t_bool		validate_elements(char *line, t_rt_list **file);
+
+// utils
+t_bool	free_return(char **target, t_bool boolean);
 double	ft_atod(char *str);
-
-//validate_utils
-int				count_elements(char **elements);
-t_is_return		is_rt_file(char *file);
-t_is_return		is_valid_type(char *type);
-t_is_return 	is_between(double conditional_num1, double conditional_num2, double num);
-t_bool			free_return(char **target, t_bool boolean);
-t_bool			validate_elements(char *line, t_rt_list **file);
-
-
-
+int		count_elements(char **elements);
