+++ src/utils/free_return.c
@@ -1,0 +1,16 @@
+#include "../../include/structures.h"
+
+t_bool	free_return(char **target, t_bool boolean)
+{
+	int	i;
+
+	i = 0;
+	while (target && target[i])
+	{
+		free(target[i]);
+		++i;
+	}
+	if (target)
+		free(target);
+	return (boolean);
+}
