+++ src/utils/split_count_utils.c
@@ -1,0 +1,9 @@
+int		count_elements(char **elements)
+{
+	int	cnt;
+
+	cnt = 0;
+	while (elements[cnt])
+		++cnt;
+	return (cnt);
+}
